Refactor: React best practices, accessibility improvements, and component decomposition
