+++ src/frontend/src/widgets/inputs/TextInputWidget/variants/SearchVariant.tsx
@@ -120,7 +120,6 @@ export const SearchVariant: React.FC<SearchVariantProps> = ({
           maxLength={props.maxLength}
           minLength={props.minLength}
           pattern={props.pattern}
-          autoFocus={props.autoFocus}
           onChange={handleChange}
           onBlur={handleBlur}
           onFocus={onFocus}
