+++ src/frontend/src/widgets/inputs/TextInputWidget/variants/TextareaVariant.tsx
@@ -39,7 +39,6 @@ export const TextareaVariant: React.FC<TextareaVariantProps> = ({
   density = Densities.Medium,
 }) => {
   const { elementRef, savePosition } = useCursorPosition(props.value, inputRef);
-
   const handleChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
     savePosition();
     onChange(e);
@@ -92,7 +91,6 @@ export const TextareaVariant: React.FC<TextareaVariantProps> = ({
           maxLength={props.maxLength}
           minLength={props.minLength}
           rows={props.rows}
-          autoFocus={props.autoFocus}
           onChange={handleChange}
           onBlur={onBlur}
           onFocus={onFocus}
